+++ ros2/src/hardware/interface/march_hardware/src/encoder/encoder.cpp
@@ -5,8 +5,7 @@
 #include "march_hardware/motor_controller/motor_controller_type.h"
 
 namespace march {
-Encoder::Encoder(size_t counts_per_rotation, MotorControllerType motor_controller_type,
-    Direction direction)
+Encoder::Encoder(size_t counts_per_rotation, MotorControllerType motor_controller_type, Direction direction)
     : total_positions_(Encoder::calculateTotalPositions(counts_per_rotation))
     , motor_controller_type_(motor_controller_type)
     , direction_(direction)
@@ -42,10 +41,9 @@ size_t Encoder::calculateTotalPositions(size_t counts_per_rotation)
 {
     if (counts_per_rotation < Encoder::MIN_COUNTS_PER_ROTATION
         || counts_per_rotation > Encoder::MAX_COUNTS_PER_ROTATION) {
-        throw error::HardwareException(
-                error::ErrorType::INVALID_ENCODER_COUNTS_PER_ROTATION,
-                "Encoder CPR (counts per rotation) of %d is not within range [%ld, %ld]",
-                counts_per_rotation, Encoder::MIN_COUNTS_PER_ROTATION, Encoder::MAX_COUNTS_PER_ROTATION);
+        throw error::HardwareException(error::ErrorType::INVALID_ENCODER_COUNTS_PER_ROTATION,
+            "Encoder CPR (counts per rotation) of %d is not within range [%ld, %ld]", counts_per_rotation,
+            Encoder::MIN_COUNTS_PER_ROTATION, Encoder::MAX_COUNTS_PER_ROTATION);
     }
     return counts_per_rotation;
 }
@@ -60,7 +58,7 @@ double Encoder::velocityIUToRadians(double velocity) const
     return velocity * getRadiansPerIU();
 }
 
-double Encoder::positionRadiansToIU(double position) const
+int32_t Encoder::positionRadiansToIU(double position) const
 {
     return position / getRadiansPerIU();
 }
